+++ R/check_balance.R
@@ -29,7 +29,7 @@
 #' target = ~ single_shift,
 #' svydesign = jvs_svy,
 #' data = admin, method_selection = "logit",
-#' control_selection = control_sel(h = 1, est_method_sel = "gee"))
+#' control_selection = control_sel(est_method = "gee", gee_h_fun = 1))
 #'
 #' ## check the balance for the standard IPW
 #' check_balance(~size, ipw_est1)